Desafios 5 ao 7; Corrige desafio1

+++ desafio1.sql
@@ -7,7 +7,7 @@ USE SpotifyClone;
 CREATE TABLE planos (
     plano_id INT PRIMARY KEY AUTO_INCREMENT,
     nome VARCHAR(20) NOT NULL,
-    valor_plano DOUBLE NOT NULL
+    valor_plano DECIMAL(5 , 2 ) NOT NULL
 )  ENGINE=INNODB;
 
 CREATE TABLE usuarios (
@@ -67,7 +67,7 @@ CREATE TABLE usuarios_cancoes (
 
 INSERT INTO planos (nome, valor_plano)
 VALUES
-  ('gratuito', 0),
+  ('gratuito', 0.00),
   ('familiar', 7.99),
   ('universitário', 5.99);
